% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pastec_api.R
\name{add_image}
\alias{add_image}
\title{Add image to Pastec index}
\usage{
add_image(image_path, image_id, server)
}
\arguments{
\item{image_path}{Path to the image file}

\item{image_id}{Integer. Unique image id.}

\item{server}{Pastec server}
}
\value{
On successful addition, silently returns list with server response.
}
\description{
Add image to Pastec index
}
\examples{
\dontrun{

erasmus1 <- system.file("img", "RP-P-1906-1485.jpg", package = "PastecR")
ps <- open_pastec_server()
str(add_image(erasmus1, 1, ps))
#> List of 3
#>  $ image_id             : int 1
#>  $ nb_features_extracted: int 1721
#>  $ type                 : chr "IMAGE_ADDED"
}
}

