% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pastec_api.R
\name{load_index}
\alias{load_index}
\title{Load an index}
\usage{
load_index(index_path, server = pastec_server())
}
\arguments{
\item{index_path}{File contianing a Pastec index.}

\item{server}{Pastec server.}
}
\value{
Returns TRUE on success, returns false (with a warning) on failure
}
\description{
Directs Pastec to load an image index from disk.
}

