% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pastec_search_results.R
\name{pastec_response}
\alias{bounding_rects}
\alias{image_ids}
\alias{pastec_response}
\alias{results_as_data_frame}
\alias{scores}
\alias{tags}
\title{Parse Pastec image search results}
\usage{
bounding_rects(pastec_response)

image_ids(pastec_response)

scores(pastec_response)

tags(pastec_response)

results_as_data_frame(pastec_response)
}
\arguments{
\item{pastec_response}{The list returned by \link{search_image}.}
}
\description{
These functions access the different components of the response to a successful Pastec image query.
}
\section{Functions}{
\itemize{
\item \code{bounding_rects}: A data.frame with the bounding box dimensions and origin for the match.

\item \code{image_ids}: A vector of returned IDs.

\item \code{scores}: A vector of similarity scores.

\item \code{tags}: A vector of tags.

\item \code{results_as_data_frame}: A data.frame with one row per search result:
\describe{
  \item{\code{height}}{integer. The height of the bounding box in the matched image.}
  \item{\code{width}}{integer. The width of the bounding box in the matched image.}
  \item{\code{x}}{integer. The origin the bounding box in the matched image.}
  \item{\code{y}}{integer. The origin of the bounding box in the matched image.}
  \item{\code{image_id}}{integer. The Pastec image IDs matched.}
  \item{\code{scores}}{integer. The image match score.}
  \item{\code{tags}}{list. A character list with any tags associated with the image.}
}
}}

