% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pastec_api.R
\name{pastec_server}
\alias{pastec_server}
\title{Link to a Pastec server}
\usage{
pastec_server(url = "localhost", port = 4212)
}
\arguments{
\item{url}{Defaults to \code{localhost://}}

\item{port}{Pastec default port is 4212}
}
\value{
Path to the Pastec server
}
\description{
Link to a Pastec server
}
\examples{
\dontrun{
ps <- pastec_server(url = "localhost", port = 4212)
add_image("image.jpg", 1, server = ps)
}
}

