% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pastec.R
\name{PastecR}
\alias{PastecR}
\title{PastecR: Wrapper for the Pastec Image Search and Recognition Engine API}
\description{
The PastecR package offers methods for interacting with the Pastec server API, as well as helper methods for interpreting the image search results in an R-friendly way.
}

