% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pastec_api.R
\name{remove_image}
\alias{remove_image}
\title{Remove image from Pastec index}
\usage{
remove_image(image_id, server)
}
\arguments{
\item{image_id}{Integer. An image id in the Pastec index.}

\item{server}{Pastec server.}
}
\value{
On successful addition, silently returns list with server response.
}
\description{
Remove a specific image from the Pastec index.
}
\examples{
\dontrun{
erasmus1 <- system.file("img", "RP-P-1906-1485.jpg", package = "PastecR")
ps <- open_pastec_server()
add_image(erasmus1, 1, ps)
str(remove_image(1, ps))
#> List of 2
#>  $ image_id: int 1
#>  $ type    : chr "IMAGE_REMOVED"
}
}

