% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pastec_api.R
\name{save_index}
\alias{save_index}
\title{Save an index}
\usage{
save_index(index_path, server = pastec_server())
}
\arguments{
\item{index_path}{File the Pastec index will be saved to. (File ending should be ".dat")}

\item{server}{Pastec server.}
}
\value{
Returns TRUE on success, returns false (with a warning) on failure
}
\description{
Directs Pastec to save its image index to a file on disk.
}

