% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pastec_api.R
\name{add_image}
\alias{add_image}
\title{Add image to Pastec index}
\usage{
add_image(image_path, image_id, server = pastec_server())
}
\arguments{
\item{image_path}{Path to the image file}

\item{image_id}{Integer. Unique image id.}

\item{server}{Pastec server}
}
\value{
On successful addition, silently returns list with the image id and response type.
}
\description{
Add image to Pastec index
}
\examples{
\dontrun{
ps <- pastec_server(url = "localhost", port = 4212)
add_image("image.jpg", 1, server = ps)
}
}

