% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pastec_api.R
\name{clear_index}
\alias{clear_index}
\title{Clear an index}
\usage{
clear_index(server)
}
\arguments{
\item{server}{Pastec server.}
}
\value{
Returns TRUE on success, returns false (with a warning) on failure
}
\description{
Directs Pastec to clear the image index.
}
\examples{
\dontrun{
ps <- pastec_server(url = "localhost", port = 4212)
clear_index(server = ps)
#> Pastec index cleared.
#> [1] TRUE
}
}

