% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pastec_servers.R
\name{hosted_pastec_server}
\alias{hosted_pastec_server}
\title{Link to a hosted Pastec server}
\usage{
hosted_pastec_server(index_id, auth_key)
}
\arguments{
\item{index_id}{Character. ID string of the index hosted at
\url{https://api.pastec.io}.}

\item{auth_key}{Character. Authentication key string for your account at
\url{https://api.pastec.io}.}
}
\value{
A list with the attributes of the server connection
}
\description{
Your \code{index_id} and \code{auth_key} can be found at
\url{https://api.pastec.io/manager/apiPage} once you make an account.
}
\seealso{
\link{open_pastec_server}
}

