% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pastec_servers.R
\name{open_pastec_server}
\alias{open_pastec_server}
\title{Link to a self-run Pastec server}
\usage{
open_pastec_server(url = "localhost", port = 4212)
}
\arguments{
\item{url}{Defaults to \code{localhost://}}

\item{port}{Pastec default port is 4212}
}
\description{
Link to a self-run Pastec server
}

