% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pastec_search_results.R
\name{pastec_response}
\alias{bounding_rects}
\alias{image_ids}
\alias{pastec_response}
\alias{results_as_data_frame}
\alias{scores}
\alias{tags}
\title{Parse Pastec image search results}
\usage{
bounding_rects(pastec_response)

image_ids(pastec_response)

scores(pastec_response)

tags(pastec_response)

results_as_data_frame(pastec_response)
}
\arguments{
\item{pastec_response}{The list returned by \link{search_image}.}
}
\description{
These functions access the different components of the response to a successful Pastec image query.
}
\section{Functions}{
\itemize{
\item \code{bounding_rects}: A data.frame with the bounding box dimensions and origin for the match.

\item \code{image_ids}: A vector of returned IDs.

\item \code{scores}: A vector of similarity scores.

\item \code{tags}: A vector of tags.

\item \code{results_as_data_frame}: A data.frame with one row per search result:
\describe{
  \item{\code{height}}{integer. The height of the bounding box in the matched image.}
  \item{\code{width}}{integer. The width of the bounding box in the matched image.}
  \item{\code{x}}{integer. The origin the bounding box in the matched image.}
  \item{\code{y}}{integer. The origin of the bounding box in the matched image.}
  \item{\code{image_id}}{integer. The Pastec image IDs matched.}
  \item{\code{scores}}{integer. The image match score.}
  \item{\code{tags}}{list. A character list with any tags associated with the image.}
}
}}
\examples{
\dontrun{
nightwatch <- system.file("img", "SK-C-5.jpg", package = "PastecR")
erasmus1 <- system.file("img", "RP-P-1906-1485.jpg", package = "PastecR")
erasmus2 <- system.file("img", "RP-P-1906-1486.jpg", package = "PastecR")
erasmus3 <- system.file("img", "RP-P-1906-1487.jpg", package = "PastecR")

ps <- open_pastec_server()
add_image(erasmus1, 1, ps)
add_image(erasmus2, 2, ps)
search_results <- search_image(erasmus1, ps)

bounding_rects(search_results)
#>   height width  x   y
#> 1    827   592 53 142
#> 2    803   567 64 159
#> 3    796   575 56 173
image_ids(search_results)
#> [1] 1 3 2
scores(search_results)
#> [1] 964  62  59
tags(search_results)
#> [1] NA NA NA
#' results_as_data_frame(search_results)
#>   height width  x   y image_id scores tags
#> 1    827   592 53 142        1    964   NA
#> 2    803   567 64 159        3     62   NA
#> 3    796   575 56 173        2     59   NA
}
}

