% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pastec_api.R
\name{remove_image}
\alias{remove_image}
\title{Remove image from Pastec index}
\usage{
remove_image(image_id, server)
}
\arguments{
\item{image_id}{Integer. An image id in the Pastec index.}

\item{server}{Pastec server.}
}
\value{
Returns TRUE on success, returns FALSE with a warning when image_id is not found.
}
\description{
Remove a specific image from the Pastec index.
}
\examples{
\dontrun{
ps <- pastec_server(url = "localhost", port = 4212)
remove_image(1, server = ps)
}
}

