% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pastec_api.R
\name{search_image}
\alias{search_image}
\title{Search Pastec index by image}
\usage{
search_image(image_path, server = pastec_server())
}
\arguments{
\item{image_path}{Path to an image to upload.}

\item{server}{Pastec server.}
}
\value{
On a successful search, returns a list generated from Pastec's JSON response.
}
\description{
Upload an image to the Pastec server to search the index for similar images.
}
\examples{
\dontrun{
ps <- pastec_server(url = "localhost", port = 4212)
search_image("image.jpg", server = ps)
}
}

